20200917 一对多插入 多对多插入

+++ src/main/java/io/github/talelin/latticy/laver/service/CategoryService.java
@@ -16,6 +16,6 @@ import java.util.Locale;
  */
 public interface CategoryService extends IService<CategoryDO> {
 
-    CategoryDO getCategoryById(Long id);
+    CategoryDO getCategoryById(Integer id);
 
 }
